Cria novos filtros, Insere bloqueios de alteração de entidades quando bloqueadas, Adiciona o campo de contagem de favoritos e duas triggers para as contagens delas.

+++ app/Services/AnnouncementService.php
@@ -352,6 +352,9 @@ class AnnouncementService implements IAnnouncementService {
     $userId         = $obAnnouncement->user_id;
     $addressId      = $obAnnouncement->address_id;
 
+    if($obAnnouncement->blocked)
+      throw new BusinessException("Não é permitido excluir um anúncio pausado", 400);
+
     (new File("user/$userId/announcement/{$id}/"))->removeAll();
     $this->obAnnouncementModel->remove($id);
     $this->obAddressService->remove($addressId);
